migration & seed ulang
