+++ models/user.js
@@ -18,7 +18,8 @@ module.exports = (sequelize, DataTypes) => {
   User.init({
     email: DataTypes.STRING,
     password: DataTypes.STRING,
-    role: DataTypes.STRING
+    role: DataTypes.STRING,
+    ProfileId: DataTypes.INTEGER
   }, {
     sequelize,
     modelName: 'User',
