+++ seeders/20240718020637-SeedCompany.js
@@ -1,0 +1,31 @@
+'use strict';
+
+/** @type {import('sequelize-cli').Migration} */
+module.exports = {
+  async up (queryInterface, Sequelize) {
+    /**
+     * Add seed commands here.
+     *
+     * Example:
+     * await queryInterface.bulkInsert('People', [{
+     *   name: 'John Doe',
+     *   isBetaMember: false
+     * }], {});
+    */
+    const data = require('../data/company.json').map(el => {
+      el.createdAt = el.updatedAt = new Date()
+      return el
+      })
+      await queryInterface.bulkInsert('Companies', data, {})
+  },
+  
+    async down (queryInterface, Sequelize) {
+      /**
+       * Add commands to revert seed here.
+       *
+       * Example:
+       * await queryInterface.bulkDelete('People', null, {});
+       */
+      await queryInterface.bulkDelete('Companies', null, {})
+    }
+  };
