+++ views/_navbar.ejs
@@ -21,7 +21,7 @@
       </div>
       <div id="navbar-collapse-with-animation" class="hs-collapse hidden overflow-hidden transition-all duration-300 basis-full grow sm:block">
         <div class="flex flex-col sm:flex-row sm:items-center sm:justify-end sm:ps-7">
-          <a class="py-3 ps-px sm:px-3 sm:py-6 font-medium text-blue-600" href="#" aria-current="page">Landing</a>
+          <a class="py-3 ps-px sm:px-3 sm:py-6 font-medium text-blue-600" href="/">Landing</a>
           <a class="py-3 ps-px sm:px-3 sm:py-6 font-medium text-gray-800 hover:text-gray-500" href="#">Account</a>
           <a class="py-3 ps-px sm:px-3 sm:py-6 font-medium text-gray-800 hover:text-gray-500" href="#">Work</a>
           <a class="py-3 ps-px sm:px-3 sm:py-6 font-medium text-gray-800 hover:text-gray-500" href="#">Blog</a>
