+++ views/landingpage.ejs
@@ -22,7 +22,7 @@
     
     <!-- heading -->
     <div class="flex items-center justify-center min-h-screen">
-        <div class="flex items-center justify-center min-h-screen">
+        <div class="flex items-center justify-center h-3/6">
             <div class="text-center max-w-4xl px-4">
                 <h1 class="mb-4 text-4xl font-extrabold leading-none tracking-tight text-gray-900 md:text-5xl lg:text-6xl dark:text-gray-800">
                     Investing in Innovation for a <span class="text-blue-600 dark:text-blue-500">Better Future</span>
